\name{mean.timeit}
\alias{mean.timeit}
\title{Calculate the mean of a 'timeit' Object}
\usage{
  \method{mean}{timeit} (x, ...)
}
\arguments{
  \item{x}{the 'timeit' object.}

  \item{...}{additional arguments supplied to
  \code{\link{mean.default}}.}
}
\description{
  This function calculates the mean running time of each
  function call.
}

