\name{plot.timeit}
\alias{plot.timeit}
\title{Plot a 'timeit' Object}
\usage{
  \method{plot}{timeit} (x, y = NULL, ...)
}
\arguments{
  \item{x}{the \code{timeit} object.}

  \item{y}{unused.}

  \item{...}{unused additional arguments.}
}
\description{
  This generates a boxplot of the timing output for a
  \code{timeit} object.
}

