\name{summary.timeit}
\alias{summary.timeit}
\title{Summarize an 'timeit' Object}
\usage{
  \method{summary}{timeit} (object, ...)
}
\arguments{
  \item{object}{an object of class \code{timeit}.}

  \item{...}{unused additional arguments.}
}
\description{
  This function generates some summary statistics for a
  \code{\link{timeit}} object.
}

